% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.arrivals.R, R/plot.attributes.R,
%   R/plot.resources.R, R/simmer.plot.R
\name{get_mon}
\alias{get_mon}
\alias{get_mon_arrivals}
\alias{get_mon}
\alias{get_mon_attributes}
\alias{get_mon}
\alias{get_mon_resources}
\alias{get_mon}
\title{Monitoring Statistics}
\usage{
get_mon_arrivals(...)

get_mon_attributes(...)

get_mon_resources(...)
}
\arguments{
\item{...}{see \code{\link[simmer]{get_mon}}.}
}
\value{
Returns a data frame of class \code{arrivals}, \code{attributes} or
\code{resources}.
}
\description{
Replacements for \code{\link[simmer:get_mon]{get_mon_arrivals}},
\code{\link[simmer:get_mon]{get_mon_attributes}} and
\code{\link[simmer:get_mon]{get_mon_resources}}.
These versions just add a new class (\code{arrivals}, \code{attributes} or
\code{resources} respectively) to the resulting data frame.
}
