% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simmer.R
\name{plot.simmer}
\alias{plot.simmer}
\title{Plot method for simmer objects}
\usage{
\method{plot}{simmer}(x, what = c("resources", "arrivals", "attributes"),
  metric = NULL, ...)
}
\arguments{
\item{x}{a single simmer environment or a list of environments representing several replications.}

\item{what}{type of plot: arrivals, attributes or resources.}

\item{metric}{specific metric. See \code{\link{plot_resources}}, \code{\link{plot_arrivals}} and
\code{\link{plot_attributes}} for more details.}

\item{...}{further arguments for each kind of plot.}
}
\description{
A method for the \code{\link{plot}} generic.
}

