% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attributes.R
\name{plot_attributes}
\alias{plot_attributes}
\title{Plot methods for attributes}
\usage{
plot(x, "attributes", metric = NULL, ...)
}
\arguments{
\item{x}{a single simmer environment or a list of environments representing several replications.}

\item{metric}{no metrics at the moment.}

\item{keys}{the keys of attributes you want to plot (if left empty, all attributes are shown).}

\item{...}{no further arguments at the moment.}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot the evolution of attributes.
}
\seealso{
\code{\link{plot.simmer}}.
}

