% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{plot_resources}
\alias{plot_resources}
\alias{plot_resources_usage}
\alias{plot_resources_utilization}
\title{Plot methods for resources}
\usage{
plot_resources(x, metric = c("usage", "utilization"), names, ...)

plot_resources_usage(x, names, items = c("system", "queue", "server"),
  steps = FALSE)

plot_resources_utilization(x, names)
}
\arguments{
\item{x}{a single simmer environment or a list of environments representing several replications.}

\item{metric}{specific metric. See \code{\link{plot_resources}}, \code{\link{plot_arrivals}} and
\code{\link{plot_attributes}} for more details.}

\item{names}{the name of the resource (character vector).}

\item{...}{further arguments for each kind of plot.}

\item{items}{the components of the resource to be plotted.}

\item{steps}{adds the changes in the resource usage.}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot
- the usage of a resource over the simulation time frame.
- the utilization of specified resources in the simulation.
}
\seealso{
\code{\link{plot_resource_utilization}},
\code{\link{plot_evolution_arrival_times}}, \code{\link{plot_attributes}}.
}

