% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.plot.R
\docType{package}
\name{simmer.plot-package}
\alias{simmer.plot-package}
\title{\pkg{simmer.plot}: Plotting Methods for \pkg{simmer}}
\description{
A set of plotting methods for \pkg{simmer} trajectories and simulations.
}
\seealso{
\pkg{simmer}'s homepage \url{http://r-simmer.org} and
GitHub repository \url{https://github.com/r-simmer/simmer.plot}.
}
\author{
Iñaki Ucar, Bart Smeets
}
