% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivals.R
\name{plot_arrivals}
\alias{plot_arrivals}
\title{Plot methods for arrivals}
\usage{
plot_arrivals(x, metric = c("activity_time", "waiting_time", "flow_time"),
  ...)
}
\arguments{
\item{x}{a single simmer environment or a list of environments representing several replications.}

\item{metric}{one of \code{c("activity_time", "waiting_time", "flow_time")}.}

\item{...}{further arguments for each kind of plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot the evolution of arrival related times:
- activity time
- waiting time
- flow time
}

