% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resources.R
\name{plot_resources}
\alias{plot_resources}
\title{Plot methods for resources}
\usage{
plot(x, "resources", metric = c("usage", "utilization"), names, ...)
}
\arguments{
\item{x}{a single simmer environment or a list of environments representing several replications.}

\item{metric}{one of \code{c("usage", "utilization")}.}

\item{names}{the name of the resource(s) (a single string or a character vector) to show.}

\item{...}{further arguments for each kind of plot:
\describe{
  \item{\code{metric = "usage"}}{
    \describe{
      \item{\code{items}}{the components of the resource to be plotted.}
      \item{\code{steps}}{adds the changes in the resource usage.}
    }
  }
}}
}
\value{
Returns a ggplot2 object.
}
\description{
Plot:
\itemize{
  \item the usage of a resource over the simulation time frame.
  \item the utilization of specified resources in the simulation.
}
}
\seealso{
\code{\link{plot.simmer}}.
}

